+++ SAGESharpTests/SAGESharpTests.csproj
@@ -61,6 +61,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="AbstractEqualityByRefTests.cs" />
+    <Compile Include="AbstractEqualityByValTests.cs" />
+    <Compile Include="MockingExtensions.cs" />
+    <Compile Include="ParameterGroup.cs" />
     <Compile Include="SLB\ASCIIExtensionsTests.cs" />
     <Compile Include="SLB\EqualityUtilsTests.cs" />
     <Compile Include="SLB\IdentifierEqualityTests.cs" />
@@ -76,10 +80,6 @@
     <Compile Include="SLB\Level\Conversation\InfoEqualityTests.cs" />
     <Compile Include="SLB\StreamExtensionsTests.cs" />
     <Compile Include="SLB\UIntExtensionsTests.cs" />
-    <Compile Include="Util\AbstractEqualityByRefTests.cs" />
-    <Compile Include="Util\AbstractEqualityByValTests.cs" />
-    <Compile Include="Util\MockingExtensions.cs" />
-    <Compile Include="Util\ParameterGroup.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
