+++ SAGESharpTests/SLB/EqualityUtilsTests.cs
@@ -1,6 +1,5 @@
 ﻿using NUnit.Framework;
 using SAGESharp.SLB;
-using SAGESharpTests.Util;
 using System.Collections.Generic;
 
 namespace SAGESharpTests.SLB
