+++ SAGESharpTests/SLB/IdentifierBinaryReaderTests.cs
@@ -1,7 +1,6 @@
 ﻿using Moq;
 using NUnit.Framework;
 using SAGESharp.SLB;
-using SAGESharpTests.Util;
 using System;
 using System.IO;
