+++ SAGESharpTests/SLB/Level/Conversation/FrameBinaryReaderTests.cs
@@ -1,7 +1,6 @@
 ﻿using Moq;
 using NUnit.Framework;
 using SAGESharp.SLB.Level.Conversation;
-using SAGESharpTests.Util;
 using System;
 using System.IO;
