+++ SAGESharpTests/SLB/Level/Conversation/InfoEqualityTests.cs
@@ -1,7 +1,6 @@
 ﻿using NUnit.Framework;
 using SAGESharp.SLB;
 using SAGESharp.SLB.Level.Conversation;
-using SAGESharpTests.Util;
 using System;
 
 using static SAGESharpTests.SLB.Level.Conversation.Defaults;
